decompose schemas, add embedding indexes
