+++ modules/odr_frontend/drizzle.config.ts
@@ -4,7 +4,7 @@ import 'dotenv/config';
 import { defineConfig } from 'drizzle-kit';
 export default defineConfig({
   out: './drizzle',
-  schema: './src/db/schema.ts',
+  schema: './src/db/schemas/index.ts',
   dialect: 'postgresql',
   dbCredentials: {
     url: "postgresql://opendatarepository:opendatarepository@localhost:35432/opendatarepository",
