+++ modules/odr_frontend/src/db/relations.ts
@@ -1,16 +1,35 @@
 /*   SPDX-License-Identifier: Apache-2.0 */
 
-import { relations } from "drizzle-orm/relations";
-import { contents, contentAuthors, users, annotationSources, annotations, teams, annotationEmbeddings, embeddingEngines, contentEmbeddings, annotationRatings, contentSets, contentEvents, contentReports, annotationReports, contentSources, annotationSourcesLink, contentSetItems, userTeams } from "./schema";
+import { relations } from 'drizzle-orm/relations';
+import {
+	contents,
+	contentAuthors,
+	users,
+	annotationSources,
+	annotations,
+	teams,
+	annotationEmbeddings,
+	embeddingEngines,
+	contentEmbeddings,
+	annotationRatings,
+	contentSets,
+	contentEvents,
+	contentReports,
+	annotationReports,
+	contentSources,
+	annotationSourcesLink,
+	contentSetItems,
+	userTeams
+} from './schemas';
 
-export const contentAuthorsRelations = relations(contentAuthors, ({one}) => ({
+export const contentAuthorsRelations = relations(contentAuthors, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentAuthors.contentId],
 		references: [contents.id]
-	}),
+	})
 }));
 
-export const contentsRelations = relations(contents, ({one, many}) => ({
+export const contentsRelations = relations(contents, ({ one, many }) => ({
 	contentAuthors: many(contentAuthors),
 	annotations: many(annotations),
 	contentEmbeddings: many(contentEmbeddings),
@@ -25,18 +44,18 @@ export const contentsRelations = relations(contents, ({one, many}) => ({
 		fields: [contents.fromUserId],
 		references: [users.id]
 	}),
-	contentSetItems: many(contentSetItems),
+	contentSetItems: many(contentSetItems)
 }));
 
-export const annotationSourcesRelations = relations(annotationSources, ({one, many}) => ({
+export const annotationSourcesRelations = relations(annotationSources, ({ one, many }) => ({
 	user: one(users, {
 		fields: [annotationSources.addedById],
 		references: [users.id]
 	}),
-	annotationSourcesLinks: many(annotationSourcesLink),
+	annotationSourcesLinks: many(annotationSourcesLink)
 }));
 
-export const usersRelations = relations(users, ({many}) => ({
+export const usersRelations = relations(users, ({ many }) => ({
 	annotationSources: many(annotationSources),
 	annotations: many(annotations),
 	annotationEmbeddings: many(annotationEmbeddings),
@@ -47,10 +66,10 @@ export const usersRelations = relations(users, ({many}) => ({
 	contentReports: many(contentReports),
 	annotationReports: many(annotationReports),
 	contents: many(contents),
-	userTeams: many(userTeams),
+	userTeams: many(userTeams)
 }));
 
-export const annotationsRelations = relations(annotations, ({one, many}) => ({
+export const annotationsRelations = relations(annotations, ({ one, many }) => ({
 	content: one(contents, {
 		fields: [annotations.contentId],
 		references: [contents.id]
@@ -66,18 +85,18 @@ export const annotationsRelations = relations(annotations, ({one, many}) => ({
 	annotationEmbeddings: many(annotationEmbeddings),
 	annotationRatings: many(annotationRatings),
 	annotationReports: many(annotationReports),
-	annotationSourcesLinks: many(annotationSourcesLink),
+	annotationSourcesLinks: many(annotationSourcesLink)
 }));
 
-export const teamsRelations = relations(teams, ({many}) => ({
+export const teamsRelations = relations(teams, ({ many }) => ({
 	annotations: many(annotations),
 	annotationEmbeddings: many(annotationEmbeddings),
 	contentEmbeddings: many(contentEmbeddings),
 	contents: many(contents),
-	userTeams: many(userTeams),
+	userTeams: many(userTeams)
 }));
 
-export const annotationEmbeddingsRelations = relations(annotationEmbeddings, ({one}) => ({
+export const annotationEmbeddingsRelations = relations(annotationEmbeddings, ({ one }) => ({
 	annotation: one(annotations, {
 		fields: [annotationEmbeddings.annotationId],
 		references: [annotations.id]
@@ -93,15 +112,15 @@ export const annotationEmbeddingsRelations = relations(annotationEmbeddings, ({o
 	user: one(users, {
 		fields: [annotationEmbeddings.fromUserId],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const embeddingEnginesRelations = relations(embeddingEngines, ({many}) => ({
+export const embeddingEnginesRelations = relations(embeddingEngines, ({ many }) => ({
 	annotationEmbeddings: many(annotationEmbeddings),
-	contentEmbeddings: many(contentEmbeddings),
+	contentEmbeddings: many(contentEmbeddings)
 }));
 
-export const contentEmbeddingsRelations = relations(contentEmbeddings, ({one}) => ({
+export const contentEmbeddingsRelations = relations(contentEmbeddings, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentEmbeddings.contentId],
 		references: [contents.id]
@@ -117,10 +136,10 @@ export const contentEmbeddingsRelations = relations(contentEmbeddings, ({one}) =
 	user: one(users, {
 		fields: [contentEmbeddings.fromUserId],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const annotationRatingsRelations = relations(annotationRatings, ({one}) => ({
+export const annotationRatingsRelations = relations(annotationRatings, ({ one }) => ({
 	annotation: one(annotations, {
 		fields: [annotationRatings.annotationId],
 		references: [annotations.id]
@@ -128,18 +147,18 @@ export const annotationRatingsRelations = relations(annotationRatings, ({one}) =
 	user: one(users, {
 		fields: [annotationRatings.ratedById],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const contentSetsRelations = relations(contentSets, ({one, many}) => ({
+export const contentSetsRelations = relations(contentSets, ({ one, many }) => ({
 	user: one(users, {
 		fields: [contentSets.createdById],
 		references: [users.id]
 	}),
-	contentSetItems: many(contentSetItems),
+	contentSetItems: many(contentSetItems)
 }));
 
-export const contentEventsRelations = relations(contentEvents, ({one}) => ({
+export const contentEventsRelations = relations(contentEvents, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentEvents.contentId],
 		references: [contents.id]
@@ -147,10 +166,10 @@ export const contentEventsRelations = relations(contentEvents, ({one}) => ({
 	user: one(users, {
 		fields: [contentEvents.setBy],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const contentReportsRelations = relations(contentReports, ({one}) => ({
+export const contentReportsRelations = relations(contentReports, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentReports.contentId],
 		references: [contents.id]
@@ -158,10 +177,10 @@ export const contentReportsRelations = relations(contentReports, ({one}) => ({
 	user: one(users, {
 		fields: [contentReports.reporterId],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const annotationReportsRelations = relations(annotationReports, ({one}) => ({
+export const annotationReportsRelations = relations(annotationReports, ({ one }) => ({
 	annotation: one(annotations, {
 		fields: [annotationReports.annotationId],
 		references: [annotations.id]
@@ -169,17 +188,17 @@ export const annotationReportsRelations = relations(annotationReports, ({one}) =
 	user: one(users, {
 		fields: [annotationReports.reportedById],
 		references: [users.id]
-	}),
+	})
 }));
 
-export const contentSourcesRelations = relations(contentSources, ({one}) => ({
+export const contentSourcesRelations = relations(contentSources, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentSources.contentId],
 		references: [contents.id]
-	}),
+	})
 }));
 
-export const annotationSourcesLinkRelations = relations(annotationSourcesLink, ({one}) => ({
+export const annotationSourcesLinkRelations = relations(annotationSourcesLink, ({ one }) => ({
 	annotation: one(annotations, {
 		fields: [annotationSourcesLink.annotationId],
 		references: [annotations.id]
@@ -187,10 +206,10 @@ export const annotationSourcesLinkRelations = relations(annotationSourcesLink, (
 	annotationSource: one(annotationSources, {
 		fields: [annotationSourcesLink.annotationSourceId],
 		references: [annotationSources.id]
-	}),
+	})
 }));
 
-export const contentSetItemsRelations = relations(contentSetItems, ({one}) => ({
+export const contentSetItemsRelations = relations(contentSetItems, ({ one }) => ({
 	content: one(contents, {
 		fields: [contentSetItems.contentId],
 		references: [contents.id]
@@ -198,10 +217,10 @@ export const contentSetItemsRelations = relations(contentSetItems, ({one}) => ({
 	contentSet: one(contentSets, {
 		fields: [contentSetItems.contentSetId],
 		references: [contentSets.id]
-	}),
+	})
 }));
 
-export const userTeamsRelations = relations(userTeams, ({one}) => ({
+export const userTeamsRelations = relations(userTeams, ({ one }) => ({
 	team: one(teams, {
 		fields: [userTeams.teamId],
 		references: [teams.id]
@@ -209,5 +228,5 @@ export const userTeamsRelations = relations(userTeams, ({one}) => ({
 	user: one(users, {
 		fields: [userTeams.userId],
 		references: [users.id]
-	}),
+	})
 }));
