+++ modules/odr_frontend/src/db/schemas/index.ts
@@ -1,0 +1,61 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+
+// Enums
+export {
+  contentsourcetype,
+  contentstatus,
+  contenttype,
+  embeddingenginetype,
+  reportstatus,
+  usertype
+} from "./enums";
+
+// Tables from annotations.ts
+export {
+  annotationSources,
+  annotations,
+  annotationEmbeddings,
+  annotationRatings,
+  annotationReports,
+  annotationSourcesLink
+} from "./annotations";
+
+// Tables from contents.ts
+export {
+  contents,
+  contentAuthors,
+  contentSources,
+  contentSets,
+  contentEvents,
+  contentReports,
+  contentSetItems
+} from "./contents";
+
+// Tables from embeddings.ts
+export {
+  embeddingEngines,
+  contentEmbeddings
+} from "./embeddings";
+
+// Tables from misc.ts
+export {
+  featureToggles,
+  verificationToken
+} from "./misc";
+
+// Tables from teams.ts
+export {
+  teams
+} from "./teams";
+
+// Tables from users.ts
+export {
+  users,
+  sessions,
+  accounts
+} from "./users";
+
+// Tables from userTeams.ts
+export {
+  userTeams
+} from "./userTeams";
